Fix ProfileResources [RUN LARGE TESTS]

+++ yt/yt/ytlib/scheduler/job_resources.cpp
@@ -307,7 +307,7 @@ void ProfileResources(
     const TJobResources& resources,
     const TString& prefix)
 {
-    #define XX(name, Name) writer->AddGauge(prefix + "/" #name, static_cast<i64>(resources.Get##Name()));
+    #define XX(name, Name) writer->AddGauge(prefix + "/" #name, static_cast<double>(resources.Get##Name()));
     ITERATE_JOB_RESOURCES(XX)
     #undef XX
 }
